refactor(installer): delegate installation to binary self-setup
