+++ crates/vp_cli_snapshots/tests/cli_snapshots/fixtures/command_self_setup/snapshots.toml
@@ -23,3 +23,19 @@ steps = [
   { argv = ["vp", "env", "setup", "--refresh"], comment = "The same upgrade handoff retries setup after the failure is repaired", snapshot = false },
   ["vpt", "stat-file", "$VP_HOME/current/bin/.vp-setup-complete", "--assert", "file"],
 ]
+
+[[case]]
+name = "command_self_setup_bootstrap_options"
+vp = "global"
+skip-platforms = ["windows"]
+steps = [
+  { argv = ["vpt", "mkdir", "-p", "external", "home", "user"], snapshot = false },
+  { argv = ["vpt", "cp", "$VP_HOME/bin/vp", "external/vp"], snapshot = false },
+  { argv = ["vpt", "chmod", "+x", "external/vp"], snapshot = false },
+  { argv = ["vpt", "write-file", "user/.bashrc", "# Existing shell configuration\n"], snapshot = false },
+  { argv = ["./external/vp"], tty = false, envs = [["HOME", "${workspace}/user"], ["VP_HOME", "${workspace}/home"], ["VP_SKIP_DEPS_INSTALL", "1"], ["VP_VERSION", "bootstrap-test"], ["VP_NODE_MANAGER", "no"], ["VP_SELF_SETUP_NO_MODIFY_PATH", "1"]], comment = "A piped standalone binary installs with the bootstrap's choices", snapshot = false },
+  ["vpt", "stat-file", "home/current/bin/.vp-setup-complete", "--assert", "file"],
+  ["vpt", "print-file", "home/config.json"],
+  ["vpt", "print-file", "user/.bashrc"],
+  ["vpt", "stat-file", "user/.zshenv", "--assert", "missing"],
+]
