+++ crates/vp_installer/Cargo.toml
@@ -15,6 +15,7 @@ path = "src/main.rs"
 clap = { workspace = true, features = ["derive"] }
 console = { workspace = true }
 indicatif = { workspace = true }
+tempfile = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
 vp_pm_cli = { workspace = true }
 vt_path = { workspace = true }
@@ -26,7 +27,6 @@ which = { workspace = true }
 winreg = { workspace = true }
 
 [dev-dependencies]
-tempfile = { workspace = true }
 vp_shared = { workspace = true, features = ["test-utils"] }
 
 [lints]
